+++ lib/src/models/genre.dart
@@ -1,0 +1,12 @@
+class Genre {
+  int id;
+  String name;
+
+  Genre.fromJson(Map<String, dynamic> map) {
+    id = map["id"];
+    name = map["name"];
+  }
+
+  static List<Genre> listFromJson(List<dynamic> list) =>
+      list.map((x) => new Genre.fromJson(x)).toList(growable: false);
+}