+++ CHANGELOG.md
@@ -9,3 +9,10 @@
 
 ## 0.0.2
 * Added methods for predefined searches: getTopRatedMovies(), getPopularMovies(), getUpcomingMovies(), getNowPlayingMovies(), getLatestMovies()
+
+## 0.0.3
+
+* Added methods for getting movie details getMovie() 
+* Partial support for response appending of getMovie(), currently only Images are appended 
+* Renamed Resolution to AssetQuality
+* Extracted Asset Path resolving methods from Configuration class to AssetResolver